vis-service: VisClient change protocol data type.
The type of taurus_cluster_data::ioctl_cmd is changed from
int to uint16_t.
Reason: it is enough to have uint16_t type to keep all available commands,
besides, this type has the predefined size.

Acked-by: Ruslan Shymkevych <ruslan_shymkevych@epam.com>

+++ meta-xt-prod-cockpit-rcar-driver-domain/recipes-extended/vis-service/files/VisClient.cpp
@@ -36,7 +36,7 @@ const int not_defined_value = std::numeric_limits<int>::max();
 
 typedef struct {
     uint64_t   value;
-    uint64_t   ioctl_cmd;
+    uint16_t   ioctl_cmd;
 }taurus_cluster_data_t;
